Switching on game-1 template
